+++ test/src/updater/blackboxInstallMac.ts
@@ -1,0 +1,7 @@
+import { getArchSuffix } from "builder-util/out/util"
+import { Arch } from "electron-builder"
+import path from "path"
+
+export function installMac(dirPath: string, arch: Arch): string {
+  return path.join(dirPath, `mac${getArchSuffix(arch)}`, `TestApp.app`, "Contents", "MacOS", "TestApp")
+}
