+++ test/src/updater/blackboxUpdateHelpers.ts
@@ -3,19 +3,20 @@ import { PM } from "app-builder-lib/out/node-module-collector"
 import { ParallelsVmManager } from "app-builder-lib/out/vm/ParallelsVm"
 import { getWindowsVm, VmManager } from "app-builder-lib/out/vm/vm"
 import { GenericServerOptions, Nullish } from "builder-util-runtime"
-import { archFromString, deepAssign, DebugLogger, getArchSuffix, log, serializeToYaml, spawn, TmpDir } from "builder-util/out/util"
-import { execFileSync, execSync } from "child_process"
-import { randomUUID } from "crypto"
+import { archFromString, deepAssign, DebugLogger, log, serializeToYaml, spawn, TmpDir } from "builder-util/out/util"
 import { Arch, Configuration, Platform } from "electron-builder"
-import { DebUpdater, PacmanUpdater, RpmUpdater } from "electron-updater"
 import { copy, existsSync, move, outputFile, readJsonSync, remove } from "fs-extra"
 import { homedir } from "os"
 import path from "path"
+import { randomUUID } from "crypto"
 import { ExpectStatic, TestContext } from "vitest"
-import { createLocalServer, getParallelsHostIP, launchAndWaitForQuit, sha256File, toVmHomePath } from "../helpers/launchAppCrossPlatform"
+import { createLocalServer, getParallelsHostIP, launchAndWaitForQuit } from "../helpers/launchAppCrossPlatform"
 import { assertPack, modifyPackageJson, PackedContext } from "../helpers/packTester"
 import { ELECTRON_VERSION } from "../helpers/testConfig"
 import { NEW_VERSION_NUMBER, OLD_VERSION_NUMBER, writeUpdateConfig } from "../helpers/updaterTestUtil"
+import { cleanupWindowsNative, installWindowsNative, installWindowsVm } from "./blackboxInstallWindows"
+import { cleanupLinux, installLinux } from "./blackboxInstallLinux"
+import { installMac } from "./blackboxInstallMac"
 
 export const optionsForFlakyE2E = { sequential: true, retry: 1 }
 
@@ -113,7 +114,6 @@ export async function doBuild(
         packed,
         packageManager: PM.PNPM,
         projectDirCreated: async (projectDir, _tmpDir, runtimeEnv) => {
-          // await outputFile(path.join(projectDir, "package-lock.json"), "{}")
           await outputFile(path.join(projectDir, ".npmrc"), "node-linker=hoisted")
 
           await modifyPackageJson(
@@ -179,7 +179,9 @@ export async function doBuild(
   }
 }
 
-export async function handleInitialInstallPerOS({
+const LINUX_TARGETS = ["AppImage", "deb", "rpm", "pacman"]
+
+async function handleInitialInstallPerOS({
   target,
   dirPath,
   arch,
@@ -192,213 +194,26 @@ export async function handleInitialInstallPerOS({
   vm?: VmManager
   perMachine?: boolean
 }): Promise<string> {
-  let appPath: string
-  if (target === "AppImage") {
-    appPath = path.join(dirPath, `TestApp.AppImage`)
-  } else if (target === "deb") {
-    DebUpdater.installWithCommandRunner(
-      "dpkg",
-      path.join(dirPath, `TestApp.deb`),
-      commandWithArgs => {
-        execSync(commandWithArgs.join(" "), { stdio: "inherit" })
-      },
-      console
-    )
-    appPath = path.join("/opt", "TestApp", "TestApp")
-  } else if (target === "rpm") {
-    RpmUpdater.installWithCommandRunner(
-      "zypper",
-      path.join(dirPath, `TestApp.rpm`),
-      commandWithArgs => {
-        execSync(commandWithArgs.join(" "), { stdio: "inherit" })
-      },
-      console
-    )
-    appPath = path.join("/opt", "TestApp", "TestApp")
-  } else if (target === "pacman") {
-    PacmanUpdater.installWithCommandRunner(
-      path.join(dirPath, `TestApp.pacman`),
-      commandWithArgs => {
-        execSync(commandWithArgs.join(" "), { stdio: "inherit" })
-      },
-      console
-    )
-    // execSync(`sudo pacman -Syyu --noconfirm`, { stdio: "inherit" })
-    // execSync(`sudo pacman -U --noconfirm "${path.join(dirPath, `TestApp.pacman`)}"`, { stdio: "inherit" })
-    appPath = path.join("/opt", "TestApp", "TestApp")
-  } else if (process.platform === "win32") {
-    // Kill any lingering NSIS installer processes left over from previous test retries.
-    // Without this, ALLOW_ONLY_ONE_INSTALLER_INSTANCE aborts the new installer (mutex conflict),
-    // and lingering mid-replacement processes cause ENOENT at the first probe launch.
-    try {
-      execSync('taskkill /F /IM "TestApp Setup.exe" /T', { stdio: "ignore" })
-      await new Promise(resolve => setTimeout(resolve, 1000))
-    } catch {
-      // no matching process — expected on the first run
-    }
-
-    // access installed app's location
-    const installBase = perMachine
-      ? (process.env["ProgramW6432"] ?? process.env["ProgramFiles"] ?? "C:\\Program Files")
-      : path.join(process.env.LOCALAPPDATA || path.join(homedir(), "AppData", "Local"), "Programs")
-    const localProgramsPath = path.join(installBase, "TestApp")
-    // this is to clear dev environment when not running on an ephemeral GH runner.
-    // Reinstallation will otherwise fail due to "uninstall" message prompt, so we must uninstall first (hence the setTimeout delay)
-    const uninstaller = path.join(localProgramsPath, "Uninstall TestApp.exe")
-    if (existsSync(uninstaller)) {
-      console.log("Uninstalling", uninstaller)
-      execFileSync(uninstaller, ["/S", "/C", "exit"], { stdio: "inherit" })
-      await new Promise(resolve => setTimeout(resolve, 5000))
-    }
-
-    const installerPath = path.join(dirPath, "TestApp Setup.exe")
-    console.log("Installing windows", installerPath)
-    // Don't use /S for silent install as we lose stdout pipe
-    execFileSync(installerPath, ["/S"], { stdio: "inherit" })
-
-    appPath = path.join(localProgramsPath, "TestApp.exe")
-  } else if (target === "nsis" && vm) {
-    // Running on macOS host with Parallels VM — install and locate app inside the VM.
-    try {
-      await vm.exec("taskkill", ["/F", "/IM", "TestApp Setup.exe", "/T"])
-      await new Promise(resolve => setTimeout(resolve, 1000))
-    } catch {
-      // no matching process — expected on the first run
-    }
-    // Query the install root once; used for both the uninstaller check and the install path.
-    const installRoot = (
-      await vm.exec("powershell.exe", [
-        "-NoProfile",
-        "-NonInteractive",
-        "-Command",
-        perMachine ? "[Environment]::GetFolderPath('ProgramFiles')" : "[Environment]::GetFolderPath('LocalApplicationData')",
-      ])
-    ).trim()
-    const installSubDir = perMachine ? "TestApp" : "Programs\\TestApp"
-    const uninstallerPath = `${installRoot}\\${installSubDir}\\Uninstall TestApp.exe`
-    try {
-      await vm.exec(uninstallerPath, ["/S", "/C", "exit"])
-      await new Promise(resolve => setTimeout(resolve, 5000))
-    } catch {
-      // no previous installation — expected on first run
-    }
-    // Use HTTP to deliver the installer: getParallelsHostIP() returns the Mac's bridge IP
-    // (e.g. 10.211.55.2) which is reachable from the VM. \\Mac\Home\ hangs on large binary reads.
-    const hostIP = getParallelsHostIP()
-    if (!hostIP) {
-      throw new Error("Cannot determine Parallels host IP for installer delivery — no prl*/bridge* interface found")
-    }
-
-    // Validate values that will be interpolated into the PowerShell script.
-    // Both are internally generated (not user input), but guard against unexpected values.
-    if (!/^[\d.]+$/.test(hostIP)) {
-      throw new Error(`Unsafe hostIP: ${hostIP}`)
-    }
-
-    // Compute the installer hash on the Mac side before serving so the Windows side can verify it.
-    const installerBinPath = path.join(dirPath, "TestApp Setup.exe")
-    const expectedSha256 = await sha256File(installerBinPath)
-    if (!/^[0-9a-f]{64}$/i.test(expectedSha256)) {
-      throw new Error(`Unexpected hash value: ${expectedSha256}`)
-    }
-
-    const { server: installerServer, port: installerPort } = await createLocalServer(dirPath, "0.0.0.0")
-    if (!Number.isInteger(installerPort) || installerPort < 1024 || installerPort > 65535) {
-      throw new Error(`Unsafe port: ${installerPort}`)
-    }
-
-    // Write the installer script to Mac home dir; the VM reads it via \\Mac\Home\ (UNC).
-    // Using -File instead of -Command avoids double-quote stripping by prlctl/cmd.exe,
-    // which allows the Add-Type heredoc and DllImport attributes to work correctly.
-    // A small PS script file (<5 KB) reads fine from \\Mac\Home\ (only large binary reads hang).
-    const scriptPath = path.join(homedir(), `.eb-nsis-${randomUUID()}.ps1`)
-    const psScript = [
-      `$tmpDir = $null`,
-      `try {`,
-      // Kill any stale installer from a previous test run — it holds the NSIS APP_GUID mutex
-      `    Stop-Process -Name 'eb-setup' -Force -ErrorAction SilentlyContinue`,
-      `    Start-Sleep -Seconds 1`,
-      `    $tmpDir = Join-Path $env:TEMP ([Guid]::NewGuid().ToString())`,
-      `    New-Item -ItemType Directory -Path $tmpDir | Out-Null`,
-      `    $dest = Join-Path $tmpDir 'eb-setup.exe'`,
-      `    Invoke-WebRequest -Uri 'http://${hostIP}:${installerPort}/TestApp%20Setup.exe' -OutFile $dest -UseBasicParsing`,
-      `    if (-not (Test-Path $dest)) { Write-Error 'Download failed'; exit 1 }`,
-      `    Write-Output ('DOWNLOAD_SIZE:' + (Get-Item $dest).Length)`,
-      // Verify download integrity; hash was computed on the Mac before the HTTP server started
-      `    $actualHash = (Get-FileHash $dest -Algorithm SHA256).Hash.ToLower()`,
-      `    $expectedHash = '${expectedSha256}'`,
-      `    if ($actualHash -ne $expectedHash) { Write-Error ('Hash mismatch: expected ' + $expectedHash + ' got ' + $actualHash); exit 1 }`,
-      `    Write-Output ('HASH_OK:true')`,
-      // Remove Mark-of-the-Web only after integrity is confirmed
-      `    Unblock-File -Path $dest -ErrorAction SilentlyContinue`,
-      `    Write-Output ('DEST_PATH:' + $dest)`,
-      `    $proc = Start-Process -FilePath $dest -ArgumentList '/S' -PassThru`,
-      `    $finished = $proc.WaitForExit(180000)`,
-      `    if (-not $finished) {`,
-      `        $proc.Kill() | Out-Null`,
-      `        Write-Error 'Installer timed out after 180s'; exit 1`,
-      `    }`,
-      `    Write-Output ('INSTALLER_EXIT:' + $proc.ExitCode)`,
-      // NSIS one-click calls quitSuccess (SetErrorLevel 0; Quit) on success → exit 0
-      `    if ($proc.ExitCode -ne 0) { Write-Error ('Installer exited with code ' + $proc.ExitCode); exit 1 }`,
-      `    Start-Sleep -Seconds 3`,
-      `    $installPath = Join-Path '${installRoot}' '${installSubDir}\\TestApp.exe'`,
-      `    Write-Output ('APP_EXISTS:' + (Test-Path $installPath))`,
-      `    if (-not (Test-Path $installPath)) { Write-Error 'App not installed at expected path'; exit 1 }`,
-      `} finally {`,
-      // PowerShell guarantees finally runs on all exit paths including exit 1
-      `    if ($tmpDir) { Remove-Item $tmpDir -Recurse -Force -ErrorAction SilentlyContinue }`,
-      `}`,
-    ].join("\n")
-
-    await outputFile(scriptPath, psScript)
-    const winScriptPath = toVmHomePath(scriptPath)
-    try {
-      const installResult = await vm.exec("powershell.exe", ["-NoProfile", "-NonInteractive", "-ExecutionPolicy", "Bypass", "-File", winScriptPath], { timeout: 300000 })
-      console.log("Installer output:", installResult)
-    } finally {
-      installerServer.close()
-      await remove(scriptPath).catch(() => {})
-    }
-    appPath = `${installRoot}\\${installSubDir}\\TestApp.exe`
-  } else if (process.platform === "darwin") {
-    appPath = path.join(dirPath, `mac${getArchSuffix(arch)}`, `TestApp.app`, "Contents", "MacOS", "TestApp")
-  } else {
-    throw new Error(`Unsupported Update test target: ${target}`)
+  if (LINUX_TARGETS.includes(target)) {
+    return installLinux(target, dirPath)
   }
-  return appPath
+  if (process.platform === "win32") {
+    return installWindowsNative(dirPath, perMachine ?? false)
+  }
+  if (target === "nsis" && vm) {
+    return installWindowsVm(dirPath, arch, vm as ParallelsVmManager, perMachine ?? false)
+  }
+  if (process.platform === "darwin") {
+    return installMac(dirPath, arch)
+  }
+  throw new Error(`Unsupported Update test target: ${target}`)
 }
 
-export async function handleCleanupPerOS({ target, perMachine }: { target: string; perMachine?: boolean }) {
-  // TODO: ignore for now, this doesn't block CI, but proper uninstall logic should be implemented
-  if (target === "deb") {
-    //   execSync("dpkg -r testapp", { stdio: "inherit" });
-  } else if (target === "rpm") {
-    // execSync(`zypper rm -y testapp`, { stdio: "inherit" })
-  } else if (target === "pacman") {
-    execSync(`pacman -R --noconfirm testapp`, { stdio: "inherit" })
-  } else if (process.platform === "win32") {
-    // Kill any lingering NSIS installer processes before running the uninstaller,
-    // so the uninstaller isn't blocked by a still-running update installer holding file locks.
-    try {
-      execSync('taskkill /F /IM "TestApp Setup.exe" /T', { stdio: "ignore" })
-      await new Promise(resolve => setTimeout(resolve, 500))
-    } catch {
-      // no matching process — ignore
-    }
-
-    // access installed app's location
-    const installBase = perMachine
-      ? (process.env["ProgramW6432"] ?? process.env["ProgramFiles"] ?? "C:\\Program Files")
-      : path.join(process.env.LOCALAPPDATA || path.join(homedir(), "AppData", "Local"), "Programs")
-    const localProgramsPath = path.join(installBase, "TestApp")
-    const uninstaller = path.join(localProgramsPath, "Uninstall TestApp.exe")
-    console.log("Uninstalling", uninstaller)
-    execFileSync(uninstaller, ["/S", "/C", "exit"], { stdio: "inherit" })
-    await new Promise(resolve => setTimeout(resolve, 5000))
-  } else if (process.platform === "darwin") {
-    // ignore, nothing to uninstall, it's running/updating out of the local `dist` directory
+async function handleCleanupPerOS({ target, perMachine }: { target: string; perMachine?: boolean }): Promise<void> {
+  if (process.platform === "win32") {
+    return cleanupWindowsNative(perMachine)
   }
+  cleanupLinux(target)
 }
 
 export async function runTestWithinServer(doTest: (rootDirectory: string, updateConfigPath: string) => Promise<void>, vm?: VmManager) {
